org-management/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
motor==3.1.1
pydantic==1.10.11
python-dotenv==1.0.0
PyJWT==2.8.0
passlib[bcrypt]==1.8.2
python-multipart==0.0.6
